Fix KeyError by checking sound existence before playing in update_display
This commit adds a check to see if the sound key is in the dictionary before trying to play the sound in the update_display() method. This prevents a KeyError from being raised when the sound is not present in the dictionary, which can happen if the mixer is not initialized.